+++ lib/misc/lwsac/lwsac.c
@@ -70,7 +70,7 @@ lwsac_get_next(struct lwsac *lac)
 }
 
 int
-lwsac_extend(struct lwsac *head, int amount)
+lwsac_extend(struct lwsac *head, size_t amount)
 {
 	struct lwsac_head *lachead;
 	struct lwsac *bf;
